Geometric morphometric investigation of Neanderthal handaxes from four archaeological units.
This is the code for my MSc Dissertation, exploring the possibility of regionalised Neanderthal cultures explored through the statistical analysis of four different handaxe groups. Two of these are from offshore collections, providing insight into the identification of such material.
(Outline data etc. soon to be uploaded).

Handaxe_GMM.R is the main code for all handaxes used in this study, providing inter- and intra-site comparison.
A240Walc_GMM.R is the code for a direct comparison of offshore material.
All other code files are for individual site analyses.